Adds logging and dev-tool support to redux store
